Overriding the equals() and hashCode() method of Feature. This is necessary because we want to check if two Features have the same float array reference as data. Currently, two Features are equal if they are the same object, or if their feature data float array have the same reference.
The FrameDropper, when replacing a Feature with the previous feature,
will create a new Feature with the feature data (a float array) of the
previous feature as its own feature data. The ID following
the ID of the previous feature is assigned to the new Feature. Overriding
the equals() method allows us to say that two Feature objects are equal
if they have the same float array *reference* as feature data,
even if they have different IDs.

The equals() method is used in GaussianMixture.getScore(Feature) to determine
if the two Features have the same array reference as the feature data.


git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@2096 94700074-3cef-4d97-a70e-9c8c206c02f5