Used IntelliJ code inspection to detect and optimize multiply and divide by 2.  The code is replaced with a shift.  Because this makes the code somewhat less clear, this optimization was only performed on "inner-loop" classes where performance matters.
This inspection reports all instances of multiplication of an integer value by a constant power of 2. These expressions may be replaced by right or left shift operations, to some possible performance improvement.
Use the check box below to enable the inspection for divisions by a power of two also. Note that replacing a power of two division by a shift does not work for negative numbers.




git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@9384 94700074-3cef-4d97-a70e-9c8c206c02f5