Adding a Clojure script example.  Here is the README
This directory contains a example Clojure script.  This example reimplements the Transciber application.

Clojure is a popular choice for Java based scripting, and is a good choice for rapid development with Sphinx4.

To try this script, install Clojure from here

    http://clojure.org/

Then cd to src/scripts/clojure and type the following to the shell

    java -cp <clojure home>clojure.jar:../../../lib/sphinx4.jar:../../../lib/jsapi.jar clojure.lang.Repl ClojureTranscriber.clj




git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@9473 94700074-3cef-4d97-a70e-9c8c206c02f5